requirements.txt
datasets
langchain
openai
langchain_community
astrapy
langchain_astradb
python-dotenv
tiktoken